+++ srcs/structures/engine/player/getters_setters/get_forward_x.c
@@ -6,13 +6,13 @@
 /*   By: ncoudsi <ncoudsi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/14 13:18:55 by ncoudsi           #+#    #+#             */
-/*   Updated: 2020/07/14 13:33:36 by ncoudsi          ###   ########.fr       */
+/*   Updated: 2020/09/02 09:56:57 by ncoudsi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cube.h"
 
-int	forward_x()
+float	forward_x()
 {
 	return (g_engine->player->forward->x);
 }
