+++ srcs/structures/engine/player/getters_setters/get_pos_y.c
@@ -6,13 +6,13 @@
 /*   By: ncoudsi <ncoudsi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/14 13:17:04 by ncoudsi           #+#    #+#             */
-/*   Updated: 2020/07/14 13:17:23 by ncoudsi          ###   ########.fr       */
+/*   Updated: 2020/09/02 09:57:37 by ncoudsi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cube.h"
 
-int	pos_y()
+float	pos_y()
 {
 	return (g_engine->player->pos->y);
 }
